Fix big amounts of duplicate code, add margins and better logging.
This commit fixes the relation between `WiGui` and `WiWidget` by making `WiGui` the subclass of `WiWidget`
This was done to reduce the amount of duplicate code related to positions, size etc.
This might be reversed later when implementing a object that is specialized in setting geometry.

Added a simple margin system

Added a log function that prints time and library name in the print statementq